[CHORE] user model 분리
